mod: trying to chase down NaN

+++ source/common.h
@@ -1,6 +1,4 @@
 #pragma once
 
-#include "Vec3.h"
-
-#define drand() (double(rand())/double(RAND_MAX))
+#define drand() double(rand())/RAND_MAX
